doubao-rag/backend/test_export/rag_package/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
